Trying to access bare ng2-fan-menu by link
I stripped the demo-site from my forked ng2-fan-menu and was working on just getting the menu compiled

Having done that, I exposed the module as a link on matrumzGPC

This commit is me trying to interface with that.

I will be taking a break to work on the mirror project.